backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
supabase>=1.0.3
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.9
sqlalchemy-utils>=0.40.0  # For SQLAlchemy result backend

# Background Tasks
celery>=5.3.4
flower>=2.0.0             # Celery monitoring

# Neo4j
neo4j>=5.14.1             # Latest version with Rust extensions support
neo4j-rust-ext>=5.14.1    # 10x faster Python driver with Rust extensions

# Authentication
python-jose>=3.3.0        # JWT
passlib>=1.7.4            # Password hashing
bcrypt>=4.0.1             # Password hashing

# Email
sendgrid>=6.10.0
jinja2>=3.1.2             # For email templates
imaplib2>=3.6             # Email retrieval

# Image Processing
pytesseract>=0.3.10
Pillow>=9.5.0

# Web Scraping
requests>=2.28.2
beautifulsoup4>=4.12.2
playwright>=1.32.1

# Payment Processing
stripe>=7.0.0

# Real-time Updates
python-socketio>=5.10.0
websockets>=12.0

# Error tracking
sentry-sdk>=1.32.0

# Testing
pytest>=7.3.1
httpx>=0.24.0

# Utilities
tenacity>=8.2.2           # Retry logic

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0 